fixing demo
