+++ public/index.html
@@ -1,10 +1,23 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-  </body>
+    <head>
+        <meta charset="utf-8" />
+    </head>
+    <style>
+        .blockly {
+            width: calc(100% - 2rem);
+            height: 40rem;
+            margin: 1rem;
+        }
+        pre {
+            border: solid 1px grey;
+            padding: 1rem;
+            margin: 1rem;
+            white-space: pre-wrap;
+        }
+    </style>
+    <body>
+        <noscript>You need to enable JavaScript to run this app.</noscript>
+        <div id="root"></div>
+    </body>
 </html>
